requirements.txt
# Core dependencies
torch>=1.13.0
transformers>=4.25.0
datasets>=2.8.0
peft>=0.6.0

# Machine Learning
scikit-learn>=1.2.0
numpy>=1.23.0
pandas>=1.5.0

# Experiment Tracking
mlflow>=2.0.0
wandb>=0.13.0

# Hyperparameter Optimization
optuna>=3.0.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.11.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
pathlib

# Optional: Accelerated training
# accelerate>=0.15.0